+++ src/Symfony/Component/Mailer/Bridge/Brevo/CHANGELOG.md
@@ -5,6 +5,8 @@ CHANGELOG
 ---
 
 * Allow configuring the SMTP port through the DSN
+* Add support for `RemoteTemplateEmail` to `BrevoApiTransport`
+* Deprecate the "templateid" and "params" email headers, use a `RemoteTemplateEmail` instead
 
 6.4
 ---