+++ src/Symfony/Component/Mime/CHANGELOG.md
@@ -9,6 +9,7 @@ CHANGELOG
  * Add `AbstractPart::setContentTypeParameter()`
  * Add the `Group` class to put a group of mailboxes in a mailbox list header (RFC 5322 and RFC 6854)
  * Add `MailboxListHeader::getAddressList()` and `MailboxListHeader::createAddressList()`
+ * Make `Email::ensureBodyValid()` protected to allow subclasses adjusting the body validation
 
 8.0
 ---